requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validación
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Variables de entorno
python-dotenv==1.0.0

# HTTP requests para APIs externas
httpx==0.26.0
aiohttp==3.9.1

# IA - Google Gemini
google-generativeai==0.3.2

# Audio processing
pydub==0.25.1
wave==0.0.2

# Utilidades
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0